feat: Create the BookController and PeopleController
